refactor: modularize Medium articles API & improve type safety
Breaking changes to Medium articles API structure:
- Split monolithic index.ts into focused modules:
  * fetchMediumFeed.ts - RSS fetching logic
  * filterArticlesByCategory.ts - Category filtering
  * mediumArticleFilterConfig.ts - Keyword rules (renamed from articleCategories)
- Remove articleCategories.ts (replaced by mediumArticleFilterConfig)
- Simplify index.ts to barrel exports only

Photo types improvements:
- Derive Photo/Category/Subcategory from database.types.ts
- Remove manual interface definitions for better type sync
- Update photos.ts to use derived types with proper type assertions
- Remove category_id from Photo type (matches updated schema)

Build tooling:
- Add dotenv-cli for secure project ID handling
- Update supabase:types script to use environment variables

This refactoring improves modularity, reduces code duplication,
and ensures types stay in sync with the database schema.
