+++ package.json
@@ -9,7 +9,7 @@
     "lint": "eslint",
     "format": "prettier --write \"**/*.{js,jsx,ts,tsx,json,css,md}\"",
     "format:check": "prettier --check \"**/*.{js,jsx,ts,tsx,json,css,md}\"",
-    "supabase:types": "supabase gen types typescript --project-id ukemvlgitweejghezgcq > src/types/database.types.ts"
+    "supabase:types": "dotenv -e .env.local -- sh -c 'supabase gen types typescript --project-id \"$SUPABASE_PROJECT_ID\" > src/types/database.types.ts'"
   },
   "dependencies": {
     "@supabase/ssr": "^0.8.0",
@@ -26,6 +26,7 @@
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "dotenv-cli": "^11.0.0",
     "eslint": "^9",
     "eslint-config-next": "15.5.4",
     "prettier": "^3.8.1",
