+++ src/lib/api/mediumArticles/utils.ts
@@ -1,6 +1,5 @@
 /**
  * Article Utilities
- *
  * Helper functions for processing Medium articles.
  */
 
@@ -52,22 +51,4 @@ export function formatDate(dateString: string): string {
     month: 'long',
     day: 'numeric',
   });
-}
-
-/**
- * Format date to relative time
- * e.g., "2 days ago", "3 weeks ago"
- */
-export function formatRelativeDate(dateString: string): string {
-  const date = new Date(dateString);
-  const now = new Date();
-  const diffInMs = now.getTime() - date.getTime();
-  const diffInDays = Math.floor(diffInMs / (1000 * 60 * 60 * 24));
-
-  if (diffInDays === 0) return 'Today';
-  if (diffInDays === 1) return 'Yesterday';
-  if (diffInDays < 7) return `${diffInDays} days ago`;
-  if (diffInDays < 30) return `${Math.floor(diffInDays / 7)} weeks ago`;
-  if (diffInDays < 365) return `${Math.floor(diffInDays / 30)} months ago`;
-  return `${Math.floor(diffInDays / 365)} years ago`;
 }
